Rename sbol to sbol2
