+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "sbol/test/SBOLTestSuite"]
-	path = sbol/test/SBOLTestSuite
+	path = sbol2/test/SBOLTestSuite
 	url = https://github.com/SynBioDex/SBOLTestSuite.git
