+++ sbol2/attachment.py
