+++ sbol2/collection.py
