+++ sbol2/component.py
