+++ sbol2/componentdefinition.py
