+++ sbol2/constants.py
