+++ sbol2/interaction.py
