+++ sbol2/model.py
