+++ sbol2/object.py
