+++ sbol2/participation.py
