+++ sbol2/property.py
