+++ sbol2/sbolerror.py
