+++ sbol2/sequence.py
