+++ sbol2/sequenceconstraint.py
