+++ sbol2/test/resources/crispr_example.xml
