+++ sbol2/test/resources/tutorial/parts.xml
