+++ sbol2/test/test_collection.py
@@ -2,16 +2,16 @@ import unittest
 
 import rdflib
 
-import sbol
+import sbol2 as sbol
 
 
-class TestImplementation(unittest.TestCase):
+class TestCollection(unittest.TestCase):
 
     def __init__(self, *args):
         super().__init__(*args)
         self.homespace = 'http://example.org'
 
-    def make_identity(self, homespace, name, version, tipe='Implementation'):
+    def make_identity(self, homespace, name, version, tipe='Collection'):
         uri = '{}/{}/{}/{}'
         return uri.format(homespace, tipe, name, version)
 
@@ -21,64 +21,61 @@ class TestImplementation(unittest.TestCase):
         sbol.Config.setOption('sbol_typed_uris', True)
         sbol.Config.setOption('sbol_compliant_uris', True)
 
-    def test_implementation_exported(self):
-        self.assertIn('Implementation', dir(sbol))
-
     def testEmptyConstructor(self):
-        # This is the default name in the implementation constructor
+        # This is the default name in the collection constructor
         name = 'example'
-        c = sbol.Implementation()
+        c = sbol.Collection()
         expected_identity = self.make_identity(self.homespace, name,
                                                sbol.VERSION_STRING)
         self.assertEqual(c.identity, rdflib.URIRef(expected_identity))
         self.assertEqual(c.displayId, rdflib.Literal(name))
         self.assertEqual(c.version, rdflib.Literal(sbol.VERSION_STRING))
-        self.assertEqual(c.rdf_type, sbol.SBOL_IMPLEMENTATION)
+        self.assertEqual(c.rdf_type, sbol.SBOL_COLLECTION)
         doc = sbol.Document()
-        doc.addImplementation(c)
-        self.assertEqual(len(doc.implementations), 1)
+        doc.addCollection(c)
+        self.assertEqual(len(doc.collections), 1)
 
     def test1ArgConstructor(self):
         name = 'foo'
-        c = sbol.Implementation(name)
+        c = sbol.Collection(name)
         expected_identity = self.make_identity(self.homespace, name,
                                                sbol.VERSION_STRING)
         self.assertEqual(c.identity, rdflib.URIRef(expected_identity))
         self.assertEqual(c.displayId, rdflib.Literal(name))
         self.assertEqual(c.version, rdflib.Literal(sbol.VERSION_STRING))
-        self.assertEqual(c.rdf_type, sbol.SBOL_IMPLEMENTATION)
+        self.assertEqual(c.rdf_type, sbol.SBOL_COLLECTION)
         doc = sbol.Document()
-        doc.addImplementation(c)
-        self.assertEqual(len(doc.implementations), 1)
+        doc.addCollection(c)
+        self.assertEqual(len(doc.collections), 1)
 
     def test2ArgConstructor(self):
         name = 'foo'
         version = '3'
-        c = sbol.Implementation(name, version)
+        c = sbol.Collection(name, version)
         expected_identity = self.make_identity(self.homespace, name,
                                                version)
         self.assertEqual(c.identity, rdflib.URIRef(expected_identity))
         self.assertEqual(c.displayId, rdflib.Literal(name))
         self.assertEqual(c.version, rdflib.Literal(version))
-        self.assertEqual(c.rdf_type, sbol.SBOL_IMPLEMENTATION)
+        self.assertEqual(c.rdf_type, sbol.SBOL_COLLECTION)
         doc = sbol.Document()
-        doc.addImplementation(c)
-        self.assertEqual(len(doc.implementations), 1)
+        doc.addCollection(c)
+        self.assertEqual(len(doc.collections), 1)
 
     def test3ArgConstructor(self):
         name = 'foo'
         version = '3'
-        rdf_type = sbol.SBOL_IMPLEMENTATION + '2'
-        c = sbol.Implementation(name, version, type_uri=rdf_type)
+        rdf_type = sbol.SBOL_COLLECTION + '2'
+        c = sbol.Collection(name, version, type_uri=rdf_type)
         expected_identity = self.make_identity(self.homespace, name,
-                                               version, 'Implementation2')
+                                               version, 'Collection2')
         self.assertEqual(c.identity, rdflib.URIRef(expected_identity))
         self.assertEqual(c.displayId, rdflib.Literal(name))
         self.assertEqual(c.version, rdflib.Literal(version))
         self.assertEqual(c.rdf_type, rdf_type)
-        # Verify that when added to a document, this implementation is
-        # not in the list of implementations. That's because the rdf_type
-        # is not SBOL_IMPLEMENTATION.
+        # Verify that when added to a document, this collection is not
+        # in the list of collections. That's because the rdf_type is
+        # not SBOL_COLLECTION.
         doc = sbol.Document()
-        doc.addImplementation(c)
-        self.assertEqual(len(doc.implementations), 0)
+        doc.addCollection(c)
+        self.assertEqual(len(doc.collections), 0)
