+++ sbol2/test/test_design.py
@@ -1,5 +1,5 @@
 import unittest
-import sbol
+import sbol2 as sbol
 
 
 class TestDesign(unittest.TestCase):
