+++ sbol2/test/test_document.py
@@ -5,7 +5,7 @@ import unittest
 
 import rdflib
 
-import sbol
+import sbol2 as sbol
 
 MODULE_LOCATION = os.path.dirname(os.path.abspath(__file__))
 TEST_LOCATION = os.path.join(MODULE_LOCATION, 'resources', 'crispr_example.xml')
