+++ sbol2/test/test_ownedobject.py
@@ -2,7 +2,7 @@ import unittest
 
 import rdflib
 
-import sbol
+import sbol2 as sbol
 
 
 class TestOwnedObject(unittest.TestCase):
