+++ sbol2/test/test_style.py
